Network: Move to standalone type
The network protocol and VMNetNetwork implementation currently are
housed on ContainerManager even though they are generally useful
types even outside of this easy to use helper type. This change moves
them to not be nested types anymore, as well as exposes a new param
on VMNetNetworks constructor so we can pass the type of network.
